move updateUserCookieConsent into UserProvider
